server/requirements.txt
category_encoders==2.6.3
dask==2024.2.0
Django==5.0.2
djangorestframework==3.14.0
ipython==8.21.0
lazypredict==0.2.12
matplotlib==3.8.2
numpy==1.26.4
pandas==2.2.0
pdfminer==20191125
plotly==5.18.0
rdkit==2023.9.4
scikit_learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
spacy==3.7.2
SpeechRecognition==3.10.1
statsmodels==0.14.1
tensorflow==2.15.0.post1
tqdm==4.66.1
